backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic
pydantic-settings
cors
bcrypt==4.0.1

